added validators to extract validation logic out of the resource classes
