+++ lib/bento/resources/emails.rb
@@ -1,6 +1,8 @@
 module Bento
   class Emails
     class << self
+      include Bento::Validators::Base
+      include Bento::Validators::EmailValidators
       # Send an email that honors subscription status
       def send(to:, from:, subject:, html_body:, personalizations: {})
         validate_email(to)
@@ -43,7 +45,6 @@ module Bento
         Bento::Response.new(response)
       end
 
-
       private
 
       def client
@@ -52,15 +53,6 @@ module Bento
 
       def default_params
         { site_uuid: Bento.config.site_uuid }
-      end
-
-      def validate_email(email)
-        raise ArgumentError, 'Email is required' if email.nil? || email.empty?
-      end
-
-      def validate_author(author)
-        raise ArgumentError, 'Author is required' if author.nil? || author.empty?
-        # Additional validation can be implemented based on system requirements
       end
     end
   end
